requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.7.0

gunicorn==20.1.0
whitenoise==6.0.0

requests==2.32.3
beautifulsoup4==4.13.4

sentence-transformers==5.4.1
chromadb==1.5.7
numpy==1.26.4

google-generativeai==0.8.3

selenium==4.28.1
webdriver-manager==4.0.2